Build route to delete notebooks
This will throw an error if a user tries to delete her last notebook
and will also reassign the user's default notebook if she deletes her
current default.

Also took the oppotunity to add an association for the default notebook.
This is not currently used, but will likely come in use later.