Bind events and listeners to switch notebooks
As part of the router refactoring, switching of notebooks is no longer
done with a link that points to a router resource. Instead, the notebook
list view has a click event for these links that updates the current
notebook and notes. Views listen to these variables so any update to
these collections/models results in various view rerenderings