Bind listener and events to switch note detail
This required redefining Nevernote.notes as a model and changing that
models attributes rather than reassigning the variable itself to point
to a different model. Otherwise, the change event wouldn't fire.

Also, the new css width on the note detail view is temporary. That
should eventually be set using jquery.